+++ ExperimentalEnemyInteractions/Script.cs
@@ -3,14 +3,9 @@ using BepInEx.Logging;
 using HarmonyLib;
 using NaturalSelection.Generics;
 using NaturalSelection.EnemyPatches;
-using UnityEngine;
 using System.Collections.Generic;
-using System.Reflection;
 using System;
 using BepInEx.Configuration;
-using System.Linq;
-using Unity.Burst.CompilerServices;
-using JetBrains.Annotations;
 using System.Text;
 using BepInEx.Bootstrap;
 using NaturalSelection.Compatibility;
@@ -18,8 +13,14 @@ using NaturalSelection.Compatibility;
 namespace NaturalSelection;
 
 [BepInPlugin(MyPluginInfo.PLUGIN_GUID, MyPluginInfo.PLUGIN_NAME, MyPluginInfo.PLUGIN_VERSION)]
+// Hard Dependencies
 [BepInDependency("fandovec03.NaturalSelectionLib", BepInDependency.DependencyFlags.HardDependency)]
 [BepInDependency("LethalNetworkAPI", BepInDependency.DependencyFlags.HardDependency)]
+// Soft Dependencies
+[BepInDependency("com.velddev.enhancedmonsters", BepInDependency.DependencyFlags.SoftDependency)]
+[BepInDependency("Entity378.sellbodies", BepInDependency.DependencyFlags.SoftDependency)]
+[BepInDependency("XuuXiaolan.ReXuvination", BepInDependency.DependencyFlags.SoftDependency)]
+[BepInDependency("BMX.LobbyCompatibility", BepInDependency.DependencyFlags.SoftDependency)]
 
 public class Script : BaseUnityPlugin
 {
@@ -51,6 +52,7 @@ public class Script : BaseUnityPlugin
     internal static bool sellBodiesPresent = false;
     internal static bool rexuvinationPresent = false;
     internal static bool CompatibilityAutoToggle = false;
+    internal static bool LobbyCompatibilityPresent = false;
 
 
     internal static Dictionary<string,bool> Bools = new Dictionary<string, bool>();
@@ -83,8 +85,9 @@ public class Script : BaseUnityPlugin
         Bools.Add(nameof(debugSpiders),debugSpiders);
         Bools.Add(nameof(debugSpiderWebs),debugSpiderWebs);
         Bools.Add(nameof(debugUnspecified),debugUnspecified);
+        CompatibilityAutoToggle = BoundingConfig.CompatibilityAutoToggle.Value;
 
-        foreach(var entry in BoundingConfig.debugEntries)
+        foreach (var entry in BoundingConfig.debugEntries)
         {
             if(Bools.ContainsKey(entry.Key))
             {
@@ -120,31 +123,41 @@ public class Script : BaseUnityPlugin
         {
             //Logger.LogInfo($"GUID in Metadata> {item.Value.Metadata.GUID} : GUID in Key> {item.Key}");
             string comment = "";
-            if (!CompatibilityAutoToggle) break;
+
+            if (!CompatibilityAutoToggle) { break; }
 
             switch (item.Key)
             {
                 case "com.velddev.enhancedmonsters":
                     {
                         enhancedMonstersPresent = true;
-                        comment = "Enhanced Monsters is present";
+                        comment = "Found Enhanced Monsters";
                         break;
                     }
                 case "Entity378.sellbodies":
                     {
                         sellBodiesPresent = true;
-                        comment = "SellbodiesFixed is present";
+                        comment = "Found SellbodiesFixed";
                         break;
                     }
                 case "XuuXiaolan.ReXuvination":
                     {
                         rexuvinationPresent = true;
-                        comment = "ReXuvination is present";
+                        comment = "Found ReXuvination";
+                        break;
+                    }
+                case "BMX.LobbyCompatibility":
+                    {
+                        LobbyCompatibilityPresent = true;
+                        comment = "Found LobbyCompatibility";
                         break;
                     }
             }
             if (comment != "") Logger.LogInfo($"{comment}. Automatically loading compatibility.");
         }
+
+        try { LobbyCompCompatibility.RegisterLobbyComp(MyPluginInfo.PLUGIN_GUID, Version.Parse(MyPluginInfo.PLUGIN_VERSION)); }
+        catch { }
 
         foreach (var item in BoundingConfig.CompatibilityEntries)
         {